+++ src/engine/cosmic_dragon_engine/cloud/rain_at.rs
@@ -1657,6 +1657,7 @@ impl super::Cloud {
             char_pool: &self.char_pool,
             previous_char_pool: &self.previous_char_pool,
             edge_fade_lut: &self.edge_fade_lut,
+            rain_shadow_lut: &self.rain_shadow_lut,
             vignette_lut: &self.vignette_lut,
             vignette_lut_cols: self.vignette_lut_dims.0,
             charset_wave_line,